make Lua scripts support optional
