+++ Source/Core/Core/PowerPC/BreakPoints.cpp
@@ -16,7 +16,6 @@
 #include "Core/Core.h"
 #include "Core/PowerPC/JitInterface.h"
 #include "Core/PowerPC/MMU.h"
-#include "Core/ScriptEngine.h"
 
 bool BreakPoints::IsAddressBreakPoint(u32 address) const
 {
@@ -275,20 +274,4 @@ bool TMemCheck::Action(Common::DebugInterface* debug_interface, u32 value, u32 a
       return true;
   }
   return false;
-}
-
-bool ScriptHooks::HasBreakPoint(u32 address) const
-{
-#if __cplusplus >= 202002L
-  return m_hooks.contains(address);
-#else
-  return m_hooks.count(address) > 0;
-#endif
-}
-
-void ScriptHooks::ExecuteBreakPoint(u32 address)
-{
-  auto range = m_hooks.equal_range(address);
-  for (auto it = range.first; it != range.second; it++)
-    it->second.Execute();
 }
