+++ Source/Core/Core/PowerPC/PPCAnalyst.cpp
@@ -200,12 +200,17 @@ static bool CanSwapAdjacentOps(const CodeOp& a, const CodeOp& b)
   int b_flags = b_info->flags;
 
   // can't reorder around breakpoints
-  if (SConfig::GetInstance().bEnableDebugging &&
-      (PowerPC::breakpoints.IsAddressBreakPoint(a.address) ||
-       PowerPC::breakpoints.IsAddressBreakPoint(b.address) ||
-       PowerPC::script_hooks.HasBreakPoint(a.address) ||
-       PowerPC::script_hooks.HasBreakPoint(b.address)))
-    return false;
+  if (SConfig::GetInstance().bEnableDebugging)
+  {
+    if (PowerPC::breakpoints.IsAddressBreakPoint(a.address) ||
+        PowerPC::breakpoints.IsAddressBreakPoint(b.address))
+      return false;
+#ifdef USE_LUA_SCRIPTS
+    if (PowerPC::script_hooks.HasBreakPoint(a.address) ||
+        PowerPC::script_hooks.HasBreakPoint(b.address))
+      return false;
+#endif
+  }
   if (b_flags & (FL_SET_CRx | FL_ENDBLOCK | FL_TIMER | FL_EVIL | FL_SET_OE))
     return false;
   if ((b_flags & (FL_RC_BIT | FL_RC_BIT_F)) && (b.inst.Rc))
