+++ Source/Core/DolphinQt/CMakeLists.txt
@@ -168,8 +168,6 @@ add_executable(dolphin-emu
   Config/PatchesWidget.h
   Config/PropertiesDialog.cpp
   Config/PropertiesDialog.h
-  Config/ScriptsWidget.cpp
-  Config/ScriptsWidget.h
   Config/SettingsWindow.cpp
   Config/SettingsWindow.h
   Config/VerifyWidget.cpp
@@ -289,6 +287,13 @@ add_executable(dolphin-emu
   Updater.h
 )
 
+if (LUA_SCRIPTS)
+  target_sources(dolphin-emu PRIVATE
+    Config/ScriptsWidget.cpp
+    Config/ScriptsWidget.h
+  )
+endif()
+
 if (NOT WIN32)
   target_sources(dolphin-emu PRIVATE
     QtUtils/SignalDaemon.cpp
@@ -336,35 +341,35 @@ if(WIN32)
 
   target_compile_options(dolphin-emu PRIVATE "-D_SILENCE_CXX17_RESULT_OF_DEPRECATION_WARNING")
 
-  # Copy Sys dir
+#Copy Sys dir
   add_custom_command(TARGET dolphin-emu POST_BUILD
     COMMAND ${CMAKE_COMMAND} -E copy_directory "${CMAKE_SOURCE_DIR}/Data/Sys" "${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/Sys"
   )
 
-  # Copy qt.conf
+#Copy qt.conf
   add_custom_command(TARGET dolphin-emu POST_BUILD
     COMMAND ${CMAKE_COMMAND} -E copy_if_different "${CMAKE_CURRENT_SOURCE_DIR}/qt.conf.win" "${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/qt.conf"
   )
 
-  # Delegate to Qt's official deployment binary on Windows to copy over the necessary Qt-specific libraries, etc.
+#Delegate to Qt's official deployment binary on Windows to copy over the necessary Qt-specific libraries, etc.
   get_target_property(MOC_EXECUTABLE_LOCATION Qt5::moc IMPORTED_LOCATION)
   get_filename_component(QT_BINARY_DIRECTORY "${MOC_EXECUTABLE_LOCATION}" DIRECTORY)
   find_program(WINDEPLOYQT_EXE windeployqt HINTS "${QT_BINARY_DIRECTORY}")
 
-  # Note: We set the PATH for the duration of this command so that the
-  #       deployment application is able to locate the Qt libraries to copy.
-  #       if the necessary paths aren't already set beforehand.
-  #
-  #       For example, consider a hypothetical emulation project named Orca.
-  #       Orca supplies its own version of Qt instead of having developers actually
-  #       install the officially supported Qt libraries -- a method that would make
-  #       wrangling around with Qt through CMake much nicer and lessen the external
-  #       library maintenance burden of the project, but alas.
-  #
-  #       In this case, as Qt is not installed through the official binary, this also
-  #       means proper path variables will not be set up, thus the need to ensure they're
-  #       always set up.
-  #
+#Note : We set the PATH for the duration of this command so that the
+#deployment application is able to locate the Qt libraries to copy.
+#if the necessary paths aren't already set beforehand.
+#
+#For example, consider a hypothetical emulation project named Orca.
+#Orca supplies its own version of Qt instead of having developers actually
+#install the officially supported Qt libraries-- a method that would make
+#wrangling around with Qt through CMake much nicer and lessen the external
+#library maintenance burden of the project, but alas.
+#
+#In this case, as Qt is not installed through the official binary, this also
+#means proper path variables will not be set up, thus the need to ensure they're
+#always set up.
+#
   add_custom_command(TARGET dolphin-emu POST_BUILD
     COMMAND "${CMAKE_COMMAND}" -E env PATH="${QT_BINARY_DIRECTORY}"
             "${WINDEPLOYQT_EXE}" --libdir="${CMAKE_RUNTIME_OUTPUT_DIRECTORY}"
@@ -376,7 +381,7 @@ if(WIN32)
   )
 endif()
 
-# Handle localization
+#Handle localization
 find_package(Gettext)
 
 if(WIN32 AND NOT Gettext_FOUND)
